tests: skip 'DefaultNetNS' test on kvm
Comparing the network namespace observed inside a VM with one observed
outside of it makes little sense; in fact, it's perfectly reasonable for
those values to be the same even if in reality they're distinct network
namespaces (due to being different kernels).

In practice, the id '4026531957' for pid 1 on both the host and in the
kernel appears fairly consistently.
This resulted in `TestNetDefaultNetNS` failing on semaphore.

I have not dug into this deeply enough to know why this issue didn't
crop up before. Presumably some kernel change on the semaphore host
resulted in it generated netns ids in a more similar way to the
coreos kernel included in stage1-kvm.
Regardless, disabling the test seems to be the correct solution to this
issue.